Add files via upload
Provide all details of your extension’s features and functionality. Include any updates.

Inventory Counter extension is developed by E-Sparks after analyzing customer purchase patterns and information that facilitate the customer in making a decision. This extension provides website visitors with real-time insights on the popularity of a product and the remaining available stock for that product. The amazing features of this extension even allow visitors to check the details of available stock and number of order for each variant as well.

With the use of this extension, admin can help visitors identify the hot selling products by displaying the number of views of those products and total orders placed for it. Admin can display the available stock for a product to save his visitors from the disappointment of finding out after ordering that item is out of stock.  

This extension adds the following additional information in the product description: Total views of the selected product, total orders placed for that particular product, and available stock of that product. Admin can also disable any of the features of extension which they do not want to include in the description. 

 

Features:
Add insights about a product in the product description
Total number of time a product was viewed 
Total number of time a product was ordered 
Total number of units available in the inventory of a product 
Insights features can be enabled/disabled
Product view information can be disabled to prevent that information from appearing in the product description. 
Total product orders information can be disabled to prevent information from appearing in the product description. 
Details of the remaining inventory units for a product can be disabled to prevent information from appearing in the product description.
Admin can enable/disable features for multiple stores differently.